requirements.txt
aniso8601==8.0.0
click==8.1.3
Flask==2.2.2
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.3
Flask-Migrate==3.1.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==2.0.1
pytz==2020.1
six==1.15.0
SQLAlchemy==1.3.18
Werkzeug==2.2.2
requests
psycopg2-binary ==2.9.5
python-dotenv
